feat(接口测试): 接口测试文档分享详情页面拆分不展示菜单
